+++ ui/src/app/core/manage-plugins/manage-plugins.service.ts
@@ -14,6 +14,7 @@ import { PluginBridgeComponent } from '@/app/core/manage-plugins/plugin-bridge/p
 import { PluginCompatibilityComponent } from '@/app/core/manage-plugins/plugin-compatibility/plugin-compatibility.component'
 import { PluginConfigComponent } from '@/app/core/manage-plugins/plugin-config/plugin-config.component'
 import { ResetAccessoriesComponent } from '@/app/core/manage-plugins/reset-accessories/reset-accessories.component'
+import { SwitchToScopedComponent } from '@/app/core/manage-plugins/switch-to-scoped/switch-to-scoped.component'
 import { UninstallPluginComponent } from '@/app/core/manage-plugins/uninstall-plugin/uninstall-plugin.component'
 import { SettingsService } from '@/app/core/settings.service'
 
@@ -261,5 +262,14 @@ export class ManagePluginsService {
     })
 
     ref.componentInstance.childBridges = childBridges
+  }
+
+  async switchToScoped(plugin: any) {
+    const ref = this.$modal.open(SwitchToScopedComponent, {
+      size: 'lg',
+      backdrop: 'static',
+    })
+
+    ref.componentInstance.plugin = plugin
   }
 }